test: mv input test to input scope
